refactor: organize widgets in primitives and components folders
