+++ src/widgets/components/editor.c
