+++ src/widgets/components/editor.h
